pyproject.toml
[tool.poetry]
name = "warehouse"
version = "0.1.0"
description = ""
authors = ["tdorosh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0.7"
djangorestframework = "3.15.2"
psycopg = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
